+++ packages/agent-toolkit/src/mcp/toolkit.ts
@@ -4,8 +4,7 @@ import { ApiClient, ApiClientConfig } from '@mondaydotcomorg/api';
 import { z } from 'zod';
 import { Tool } from '../core/tool';
 import { allGraphqlApiTools, allMondayAppsTools } from '../core/tools';
-import { filterApiTools } from '../core/tools/platform-api-tools/utils';
-import { filterMondayAppsTools } from 'src/core/tools/monday-apps-tools/utils';
+import { filterApiTools, filterMondayAppsTools } from '../utils';
 
 export type ToolsConfiguration = {
   include?: string[];
