+++ packages/agent-toolkit/src/utils/tools/tools-filtering.utils.ts
@@ -1,7 +1,7 @@
 import { ApiClient } from '@mondaydotcomorg/api';
-import { ToolsConfiguration } from 'src/mcp/toolkit';
-import { ToolSubType } from '../../tool';
-import { BaseMondayApiTool } from './base-monday-api-tool';
+import { Tool, ToolSubType } from 'src/core';
+import { BaseMondayApiTool } from 'src/core/';
+import { ToolsConfiguration } from 'src/mcp';
 
 export function filterApiTools<T extends new (api: ApiClient) => BaseMondayApiTool<any>>(
   tools: T[],
@@ -38,6 +38,42 @@ export function filterApiTools<T extends new (api: ApiClient) => BaseMondayApiTo
   if (config.readOnlyMode) {
     filteredTools = filteredTools.filter((tool) => {
       const toolInstance = new tool(apiClient);
+      return toolInstance.subType === ToolSubType.READ;
+    });
+  }
+
+  return filteredTools;
+}
+
+export function filterMondayAppsTools<T extends new (...args: any[]) => Tool<any, any>>(
+  tools: T[],
+  mondayApiToken: string,
+  config?: ToolsConfiguration,
+): T[] {
+  if (!config) {
+    return tools;
+  }
+
+  if (!config.enableMondayAppsTools) {
+    return [];
+  }
+  let filteredTools = tools;
+
+  if (config.include) {
+    filteredTools = tools.filter((tool) => {
+      const toolInstance = new tool(mondayApiToken);
+      return config.include?.includes(toolInstance.name);
+    });
+  } else if (config.exclude) {
+    filteredTools = tools.filter((tool) => {
+      const toolInstance = new tool(mondayApiToken);
+      return !config.exclude?.includes(toolInstance.name);
+    });
+  }
+
+  if (config.readOnlyMode) {
+    filteredTools = filteredTools.filter((tool) => {
+      const toolInstance = new tool(mondayApiToken);
       return toolInstance.subType === ToolSubType.READ;
     });
   }
